demo/images/src/aramis/requirements.txt
Flask==0.12.5
grpcio==1.28.1
grpcio-tools==1.28.1
bitstruct==8.10.0
pyinstaller==3.6
